pyproject.toml
[project]
name = "weldfatigue"
version = "0.1.0"
description = "Fatigue and shock resistance validation for welded automobile metal assemblies"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [{name = "Reda B."}]

dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.1",
    "pydantic>=2.5",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "fpdf2>=2.8",
    "jinja2>=3.1",
    "streamlit>=1.30",
    "loguru>=0.7",
]

[project.optional-dependencies]
solvers = [
    "lasso-python>=1.5",
    "pyNastran>=1.4",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]
notebooks = [
    "jupyter>=1.0",
    "jupyterlab>=4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
    "benchmark: IIW benchmark validation tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
